add empty .bbappend files
Now most recipes used in Tizen have at least an empty .bbappend file
in the right directory. This will simplify the next steps because one
does not have to locate the upstream directory for each recipe
one-by-one. Instead a find command for <recipe>_%.bbappend in
meta-tizen is enough.

This files will not stay empty for long: at least SECTION information
needs to be added for every single one.

Change-Id: I5f88c0106bce943b9c6f6ee653a4a88fe5a3997b

Signed-off-by: Ronan Le Martret <ronan@fridu.net>